requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi[all]
h11==0.14.0
httptools==0.6.1
idna==3.10
pydantic==2.9.2
pydantic-settings==2.4.0
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.1
